src/requirements.txt
azure-ai-projects>=1.0.0b5
azure-identity>=1.15.0
openpyxl>=3.1.2
xlrd>=2.0.1
streamlit>=1.40.0
azure-ai-textanalytics>=5.3.0
pandas>=2.0.0
# Optional: only needed for standalone sentiment_agent.py (direct SDK path)
# azure-ai-textanalytics>=5.3.0
# Optional: only needed for Fabric OneLake data source
# azure-storage-file-datalake>=12.14.0
